Land row 10 (ph56): the proof costs exactly nothing, and the row bought it
Gate PASS, failures [], complete_run true, contract e76ae727f70b,
blocked_rows null, identity norel as pinned. Brackets 66/0 unmoved and
20/0 -> 22/0, the +2 being exactly this row's two records, re-run by me.
quota: 10 built, 30 owed, T5 at 2.

R4 -> R5 IS 0.000 Ir/call AND 0.000 % ON BOTH INPUTS, identity norel, and
family B reads 0.00 across all 32 pad residues. But ph55 shipped +0.071 %
on the same step, and the difference is a representation choice: Verus
models usize as 32 OR 64 bits and will not assume a u64->usize cast is
lossless, so R5 must take the modulus before the cast. ph55 resolved that
with two spellings; ph56 normalised all four Rust rungs onto one, nine
expressions across three files, so R4 and R5 are one exec text and the
R2->R3 and R3->R4 gradients carry no representation change. So "the proof
is free" is a claim about how the ladder was WRITTEN, not only about
Verus, and this is the first row to make it exactly zero by construction
rather than approximately zero by luck.

THE CROSS-LANGUAGE SIGN FLIPS BETWEEN THE TWO C COMPILERS INSIDE ONE
STATISTIC: R4/R5 against the R1h C cells, W1, O3/isolated, large.bin, is
-8.423 % against c-gcc-h and +3.811 % against c-clang-h. The row publishes
both columns and both signs rather than picking the flattering one -- F108
working on the first row built after it landed, which is the strongest
evidence that rule could have. inside_share predicts it, and
controls/statistic.py checks the wide and narrow halves in opposite
directions so neither claim can rot silently.

The gate deleted two of three preconditions this row copied from ph55.
req-mut found 16 <= len and len <= 8 * MAX_STMT not load-bearing and both
are gone: ph56 ESTABLISHES in code what ph55 ASSUMES in a signature. A
precondition that is not load-bearing narrows the admissible call sites
for nothing, and reading would not have caught it.

The mutation test is the argument: five mutants, all fail, and the one
that matters deletes the upstream fix's three lines and stops the file
verifying at the assert that re-establishes the obligation. pristine
re-verifies at 66/0 on every run so a broken toolchain cannot read as a
clean sweep.

And an rlimit error is a symptom, not a size. The floor bisects to 2, the
same as ph55, but the first draft's overrun was a wrong loop invariant --
two clauses that cannot hold at a break declared invariant instead of
invariant_except_break. Raising the limit would have shipped it.

Predictions: 1 upheld (10/10 twins, none justified away, so F97's
narrowing survives a second row); 2 upheld on mechanism, with the guard at
1.06-1.55 Ir per emitted statement flat across a 4x change in statement
count and the percentage stated as a loose upper bound on PHP's rather
than PHP's; 3 honoured, no endpoint prediction made.

Declared gap: the R3->R4 step is not decomposed. That is an endpoint
search, a separate task on every row.

Box 20/20, findings 112, layer 0 stale, all checkers green except
citecheck on item 97's known debt.

+++ patterns-php/ph56-fetchmode-arith/safe_naive.rs
@@ -242,7 +242,7 @@ fn fetch_dimension_address(
     }
     if slots[b].ty == IS_ARRAY {
         // :930
-        let a = (slots[b].lval as usize) % NVAR;
+        let a = (slots[b].lval % (NVAR as u64)) as usize;
         if op2_type == T_UNUSED {
             // :935 -- THE APPEND
             let n = anext[a];
@@ -440,14 +440,14 @@ pub fn kernel(buf: &[u8], off: usize, len: usize) -> u64 {
                 let mut result: u64 = 0;
                 if slots[c].ty == IS_ARRAY {
                     // :3967
-                    let a = (slots[c].lval as usize) % NVAR;
+                    let a = (slots[c].lval % (NVAR as u64)) as usize;
                     // ⛔⛔⛔ :3973 -- `switch (offset->type)`. In C `offset` is
                     // NULL whenever op2 is IS_UNUSED and this is a read of
                     // address 0x14. In safe Rust the Option cannot be opened
                     // without this call, so the SAME wrong program panics.
                     let off = offset.unwrap();
                     if slots[off].ty == IS_LONG {
-                        let idx = (slots[off].lval as usize) % DIM;
+                        let idx = (slots[off].lval % (DIM as u64)) as usize;
                         let s = ARR_BASE + a * DIM + idx;
                         if idx < anext[a] as usize {
                             isset = 1;
